Réglage de bug
- Les joueurs ne changent plus quand on appui sur les flèches
- Les joueurs ne bougnent plus comme des ondes lumineuses (genre
vraiment vite)

+++ Travail Final/Travail Final/CMap.h
@@ -73,7 +73,7 @@ public:
 			strTampo = chrTmp[20];
 			strTampo += chrTmp[21];
 			strTampo += chrTmp[22];
-			m_VecteurGravite = new CVecteur2D((float)SDL_atof(strTampo.c_str()), 90.0);
+			m_VecteurGravite = new CVecteur2D(SDL_atof(strTampo.c_str()), 90.0f);
 			FichierMap.getline(chrTmp, 50);
 			
 			FichierMap.getline(chrTmp, 50);
